Now renamed the new modules to remove the NEW_ prefix and replace the old modules.
